backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pillow
pytesseract